KissABC

KissABC.Factored

Base.rand

Distributions.pdf

KissABC.ABC

KissABC.ABCDE

KissABC.ABCSMCPR

KissABC.compute_kernel_scales

KissABC.deperturb

KissABC.kernel

KissABC.kerneldensity

KissABC.perturb

KissABC.Factored — Type

Factored{N} <: Distribution{MixedVariate, MixedSupport}

a Distribution type that can be used to combine multiple Distribution's and sample from them.

Example: it can be used as prior = Factored(Normal(0,1), Uniform(-1,1))

KissABC.ABC — Function

ABC(prior, simulation, data, distance, α_target; nparticles = 100, params = (), parallel = false)

Classical ABC rejection algorithm.

Arguments:

prior: a Distribution to use for sampling candidate parameters

simulation: simulation function sim(prior_sample, constants) -> data that accepts a prior sample and the params constant and returns a simulated dataset

data: target dataset which must be compared with simulated datasets

distance: distance function dist(x,y) that return the distance (a scalar value) between x and y

α_target: target acceptance rate for ABC rejection algorithm

nparticles: number of samples from the approximate posterior that will be returned

params: an optional set of constants to be passed as second argument to the simulation function

parallel: when set to true multithreaded parallelism is enabled

KissABC.ABCDE — Function

ABCDE(prior, simulation, data, distance, ϵ_target; nparticles = 100, maxsimpp = 1000, parallel = false, params = (), verbose = true)

A sequential monte carlo algorithm inspired by differential evolution, work in progress, very efficient (inspired by Turner et al. 2011, as this implementation is not 1 to 1)

Arguments:

prior: a Distribution to use for sampling candidate parameters

simulation: simulation function sim(prior_sample, constants) -> data that accepts a prior sample and the params constant and returns a simulated dataset

data: target dataset which must be compared with simulated datasets

distance: distance function dist(x,y) that return the distance (a scalar value) between x and y

ϵ_target: maximum acceptable distance between simulated datasets and the target dataset

nparticles: number of samples from the approximate posterior that will be returned

maxsimpp: average maximum number of simulations per particle

parallel: when set to true multithreaded parallelism is enabled

params: an optional set of constants to be passed as second argument to the simulation function

verbose: when set to true verbosity is enabled

KissABC.ABCSMCPR — Function

ABCSMCPR(prior, simulation, data, distance, ϵ_target; nparticles = 100, maxsimpp = 1000.0, α = 0.3, c = 0.01, parallel = false, params = (), verbose = true)

Sequential Monte Carlo algorithm (Drovandi et al. 2011).

Arguments:

prior: a Distribution to use for sampling candidate parameters

simulation: simulation function sim(prior_sample, constants) -> data that accepts a prior sample and the params constant and returns a simulated dataset

data: target dataset which must be compared with simulated datasets

distance: distance function dist(x,y) that return the distance (a scalar value) between x and y

ϵ_target: maximum acceptable distance between simulated datasets and the target dataset

nparticles: number of samples from the approximate posterior that will be returned

maxsimpp: average maximum number of simulations per particle

α: proportion of particles to retain at every iteration of SMC, other particles are resampled

c: probability that a sample will not be updated during one iteration of SMC

parallel: when set to true multithreaded parallelism is enabled

params: an optional set of constants to be passed as second argument to the simulation function

verbose: when set to true verbosity is enabled

Base.rand — Method

rand(rng::AbstractRNG, factoreddist::Factored)

function to sample one element from a Factored object

Distributions.pdf — Method

pdf(d::Factored, x) = begin

Function to evaluate the pdf of a Factored distribution object

KissABC.compute_kernel_scales — Function

compute_kernel_scales(prior::Distribution, V)

Function whose purpose is to compute the characteristic scale of the perturbation kernel appropriate for prior given the Vector V of parameters

KissABC.deperturb — Function

deperturb(sample::T, r1::T, r2::T)

Differential Evolution perturbation kernel, generalizing this interface is WIP

Arguments:

sample

r1

r2

KissABC.kernel — Function

kernel(prior::Distribution, c, scale)

Function whose purpose is returning the appropriate Distribution to use as a perturbation kernel on sample c and characteristic scale

Arguments:

prior: prior distribution

c: sample acting as center of perturbation kernel

scale: characteristic scale of perturbation kernel

KissABC.kerneldensity — Function

kerneldensity(prior::Distribution, scales, s1, s2)

Function whose purpose is returning the probability density of observing s2 under the kernel centered on s1 with scales given by scales and appropriate for prior.

KissABC.perturb — Function

perturb(prior::Distribution, scales, sample)

Function whose purpose is perturbing sample according to the appropriate kernel for prior with characteristic scales.