Reference

KissABC.ApproxKernelizedPosterior

KissABC.Factored

Base.length

Base.rand

Distributions.logpdf

Distributions.pdf

KissABC.cdf_g_inv

KissABC.mcmc

KissABC.sample_g

KissABC.ApproxKernelizedPosterior — Type

ApproxPosterior(
    prior::Distribution,
    cost::Function,
    max_cost::Real
)

this function will return a type which can be used in the mcmc function as an ABC density, this type works by assuming uniformly distributed errors in [-ϵ,ϵ], ϵ is specified in the variable max_cost.

KissABC.Factored — Type

Factored{N} <: Distribution{Multivariate, MixedSupport}

a Distribution type that can be used to combine multiple UnivariateDistribution's and sample from them. Example: it can be used as prior = Factored(Normal(0,1), Uniform(-1,1))

Base.length — Method

length(p::Factored) = begin

returns the number of distributions contained in p.

Base.rand — Method

rand(rng::AbstractRNG, factoreddist::Factored)

function to sample one element from a Factored object

Distributions.logpdf — Method

logpdf(d::Factored, x) = begin

Function to evaluate the logpdf of a Factored distribution object

Distributions.pdf — Method

pdf(d::Factored, x) = begin

Function to evaluate the pdf of a Factored distribution object

KissABC.cdf_g_inv — Method

Inverse cdf of g-pdf, see eq. 10 of Foreman-Mackey et al. 2013.

KissABC.mcmc — Method

function mcmc(
    density::AbstractApproxDensity;
    nparticles::Int,
    generations,
    rng::AbstractRNG = Random.GLOBAL_RNG,
    parallel = false,
    verbose = 2,
    bare_particles=false
)

This function will run an Affine Invariant MC sampler on the ABC density defined in density, the ensemble will contain nparticles particles, and each particle will evolve for a total number of steps equal to generations.

KissABC.sample_g — Method

Sample from g using inverse transform sampling. a=2.0 is recommended.